+++ src/app/(main)/profile/setting/edit/components/ProfileEdit.client.tsx
@@ -3,6 +3,7 @@ import EditProvider from './EditProvider.client';
 import Step1 from './step1/Step1.client';
 import Step2 from './step2/Step2.client';
 import { useGetProfile } from '@/apis/profile';
+import { ModalProvider } from '@/hooks/useModal';
 import { useState } from 'react';
 
 export default function ProfileEdit() {
@@ -11,8 +12,10 @@ export default function ProfileEdit() {
 
   return (
     <EditProvider defaultValues={defaultProfileData}>
-      {step === 1 && <Step1 onNext={() => setStep(2)} />}
-      {step === 2 && <Step2 onPrev={() => setStep(1)} />}
+      <ModalProvider>
+        {step === 1 && <Step1 onNext={() => setStep(2)} />}
+        {step === 2 && <Step2 onPrev={() => setStep(1)} />}
+      </ModalProvider>
     </EditProvider>
   );
 }
