+++ src/components/Modal/index.ts
@@ -1,5 +1,5 @@
 export { default as ModalWrapper } from './ModalWrapper.client';
-export { default as BirthdayBottomSheet } from './BirthdayBottomSheet.client';
+export { default as BirthdayBottomSheet } from '../../app/(sub)/join/funnels/step4/components/BirthdayBottomSheet.client';
 export { default as Toast } from './Toast.client';
 export { default as BottomSheet } from './BottomSheet';
 export * from './Modal.client';
